dom/file/tests/test_fileapi_slice_memFile_2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for File API + Slice (in memory)</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="common_blob.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>

<p id="display">
  <canvas id=canvas width=1100 height=1100 hidden moz-opaque></canvas>
</p>

<script type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.requestLongerTimeout(4);

createCanvasURL()
.then(data => {
  let cx = $("canvas").getContext('2d');
  let memFile = cx.canvas.mozGetAsFile("image/png");
  return testSlice(memFile, data.length, "image/png", data, "memFile", RANGE_2);
})

.then(SimpleTest.finish);

</script>
</body>
</html>
